QasBot/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
chardet==3.0.4
future==0.16.0
idna==2.5
python-telegram-bot==6.1.0
requests==2.18.1
urllib3==1.21.1
